refactor: move things into event module
